+++ react/components/AccountOptions.js
@@ -1,0 +1,12 @@
+
+import React, { Component } from 'react'
+
+// Component that shows account options to the user.
+class AccountOptions extends Component {
+    render() {
+        return (
+            <h4>Bem vindo!</h4>
+        )
+    }
+}
+export default AccountOptions